move circom frontend related structs into frontend/circom dir
